+++ look/example/lib/theme.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+ThemeData lightTheme() => ThemeData(
+      primarySwatch: Colors.blue,
+      brightness: Brightness.light,
+    );
+
+ThemeData darkTheme() => ThemeData(
+      primarySwatch: Colors.amber,
+      brightness: Brightness.dark,
+    );