+++ look_generator/lib/look_generator.dart
@@ -1,7 +1,3 @@
-library look_generator;
-
-//import 'package:analyzer/dart/constant/value.dart';
-
 import 'package:analyzer/dart/element/element.dart';
 import 'package:code_builder/code_builder.dart';
 import 'package:dart_style/dart_style.dart';